Fix Gradle version & locking, add auto-config to minimize init requirements.
