+++ src/main/java/bio/terra/common/logging/README.md
@@ -11,25 +11,55 @@ before it gets relayed to Cloud Logging for storage. See
 [Structured Logging](https://cloud.google.com/logging/docs/structured-logging#special-payload-fields) docs
 and the [GoogleJsonLayout.java](GoogleJsonLayout.java) file for more details.
 
-To support local development, JSON-formatted logging will be turned off when the
-`human-readable-logging` Spring profile is active. 
-
 ## Quickstart
 
-Typical logging setup for a Terra Spring application:
+No explicit configuration is needed. The `LoggingAutoConfiguration` class will
+be loaded by Spring if this package is on the classpath.
+
+In order to initialize the logging config as early as possible, it is 
+recommended to add the initializer directly to the Spring application:
 
 ```
-// Scan the logging package for annotated components
-@SpringBootApplication(scanBasePackages = {"bio.terra.common.logging", ...})
-public class MyApp {
   public static void main(String[] args) throws Exception {
-    SpringApplication app = new SpringApplication(MyApp.class);
-    // Insert the logging initializer into the app startup chain.
-    app.addInitializers(new LoggingInitializer());
-    app.run(args);
+    new SpringApplicationBuilder(MyApplication.class)
+        .initializers(new LoggingInitializer())
+        .run(args);
   }
-}
 ```
+
+### Human-readable logging
+
+To support local development, JSON-formatted logging will be turned off when the
+`human-readable-logging` Spring profile is active.
+
+You can add this into a Gradle task and/or include it on the command-line:
+
+```
+# Command-line arguments
+./gradlew bootRun --args='--spring.profiles.active=human-readable-logging'
+
+# Gradle task wrapper
+task bootRunDev {
+    bootRun.configure {
+        systemProperty "spring.profiles.active", 'human-readable-logging'
+    }
+}
+bootRunDev.finalizedBy bootRun
+```
+
+### Turning off auto-config
+
+To disable all auto-configuration from this package, exclude the 
+`LoggingAutoConfiguration` class with an app-level annotation:
+
+```
+@SpringBootApplication(
+    exclude = {
+      LoggingAutoConfiguration.class
+    })
+```
+
+## Structured logs
 
 To log structured data to Cloud Logging:
 
@@ -38,3 +68,5 @@ log.info("My message", LoggingUtils.jsonFromString("{eventType: 'very-rare-event
 
 log.info("My message", LoggingUtils.structuredLogData("event", myEventObject));
 ```
+
+See [LoggingUtils](LoggingUtils.java) Javadoc for more details.
