+++ src/test/java/bio/terra/common/logging/LoggingTest.java
@@ -20,6 +20,7 @@ import javax.servlet.ServletResponse;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.jupiter.api.Tag;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -63,6 +64,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 // Refer to an external properties file to define a Spring application name and version, which is
 // included in the JSON output if available.
 @ActiveProfiles("logging-test")
+@Tag("unit")
 public class LoggingTest {
 
   @Autowired private TestRestTemplate testRestTemplate;
